BoardNode and Heuristic cpps
BoardNode is a class describing a node on the tree of possible board
configurations. Each node has a parent(except for the first node), and
up to 6 children.

Heuristic provides a method to look at the board state (a vector of 14
ints), and return a double fitness score describing whether the board is
good for us or bad for us.

Heuristic has "balancer" variables which are multiplied against
individual portions of the score, we can modify the balancers to tweak
which heuristic is more valuable.